2024-08-19 Roman (stable) - !:SYS,DAQ.{DAQGate,JavaLikeCalc,OPC_UA.libOPC_UA,System},Protocol.ModBus,Transport.         SSL,UI.{QTCfg,QTStarter,Vision,WebCfgD,WebVision}: The source code some cleaned from building warnings. - SYS Type of the limit variables limUserFile_SZ, limUserIts_N changed to unsigned.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2985 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/JavaLikeCalc/freefunc.cpp
@@ -3009,7 +3009,7 @@ void Func::cntrCmdProc( XMLNode *opt )
 
 		    XMLNode *nd = ctrMkNode("fld", opt, -1, ("/exec/io/"+io(iIO)->id()).c_str(), nprm1,
 						((io(iIO)->flg()&IO::Return)?R_R_R_:RWRW__), "root", grp);
-		    if(nd && nOff < nprm.size()) nd->setAttr("help",nprm.substr(nOff));
+		    if(nd && nOff < (int)nprm.size()) nd->setAttr("help",nprm.substr(nOff));
 		}
 	return;
     }
